Remove more unnecessary files
