+++ Intake/AllergyView/AllergyLLMAssistant.swift
