+++ Intake/AllergyView/AllergyRecords.swift
