+++ Intake/AllergyView/ReactionSectionView.swift
