+++ Intake/AllergyView/ReactionView.swift
