+++ Intake/ChiefComplaintView/LLMInteraction.swift
