+++ Intake/MedicationView/MedicationLLMAssistant.swift
