+++ Intake/SurgeryView/SurgeryLLMAssistant.swift
